--- src/main/java/com/programmers/springbootjpa/domain/customer/CustomerRepository.java
@@ -1,8 +1,0 @@
-package com.programmers.springbootjpa.domain.customer;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
-}
