+++ src/main/java/com/programmers/springbootjpa/domain/mission1/customer/CustomerRepository.java
@@ -1,0 +1,8 @@
+package com.programmers.springbootjpa.domain.mission1.customer;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
+}
