+++ src/main/java/com/programmers/springbootjpa/domain/mission3/item/ItemRepository.java
@@ -1,0 +1,8 @@
+package com.programmers.springbootjpa.domain.mission3.item;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ItemRepository extends JpaRepository<Item, Long> {
+}
