+++ src/main/java/com/programmers/springbootjpa/domain/mission3/member/MemberRepository.java
@@ -1,0 +1,8 @@
+package com.programmers.springbootjpa.domain.mission3.member;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
