+++ src/main/java/com/programmers/springbootjpa/domain/mission3/order/OrderRepository.java
@@ -1,0 +1,8 @@
+package com.programmers.springbootjpa.domain.mission3.order;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface OrderRepository extends JpaRepository<Order, Long> {
+}
