+++ src/main/java/com/programmers/springbootjpa/domain/mission3/order/OrderStatus.java
@@ -1,0 +1,6 @@
+package com.programmers.springbootjpa.domain.mission3.order;
+
+public enum OrderStatus {
+    OPENED,
+    CANCELLED
+}
