+++ src/main/java/com/programmers/springbootjpa/domain/mission3/orderItem/OrderItemRepository.java
@@ -1,0 +1,8 @@
+package com.programmers.springbootjpa.domain.mission3.orderItem;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
+}
