--- src/main/java/com/programmers/springbootjpa/domain/order/BaseEntity.java
@@ -1,20 +1,0 @@
-package com.programmers.springbootjpa.domain.order;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@EntityListeners(value = {AuditingEntityListener.class})
-@MappedSuperclass
-public abstract class BaseEntity {
-
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
-}
