--- src/main/java/com/programmers/springbootjpa/domain/order/ItemRepository.java
@@ -1,8 +1,0 @@
-package com.programmers.springbootjpa.domain.order;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ItemRepository extends JpaRepository<Item, Long> {
-}
