--- src/main/java/com/programmers/springbootjpa/domain/order/OrderStatus.java
@@ -1,6 +1,0 @@
-package com.programmers.springbootjpa.domain.order;
-
-public enum OrderStatus {
-    OPENED,
-    CANCELLED
-}
